created tests_booking.py
